+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - output class probability for every gene/protein
 - simplify the calculation of cluster stats
 - class/cluster numbering starts at 1 (0-based in autoclass output)
+- add dendrogram of classes
 
 **0.1.12**
 - update write_cluster_stats() methods with mean/std column